port: Mojo 1.0.0rc0 + max 26.5.0rc0 — FFI proxy, max.gpu moves, Int32 kernel args
- New max==26.5.0rc0 dep (pixi.toml + recipe.yaml run-reqs): std.gpu.host/
  memory, barrier and block primitives moved to the max Mojo package.
- SimdjsonFFI: get_function's origin-bound callable can't be stored beside
  its handle — 25 fn-pointer fields became _CFn[R] proxies resolved via
  get_symbol (same dance as stdlib _DLCallable; struct owns the handle).
- GPU kernels take UInt32 scalar args (Int/UInt aren't DevicePassable);
  read-only fused-kernel input is ImmutAnyOrigin (imm borrows map mut=False).
- trait_downcast retired for refinement (conforms_to guard / comptime
  assert) in reflection ser/de.
- Deprecation sweep to zero warnings (unsafe_memcpy/load, unsafe_offset
  subscripts, Pointer, __deinit__); perf_counter Int; String literal
  indexing via as_bytes.

tests-cpu + tests-gpu green, zero warnings.

+++ json/gpu/parser.mojo
@@ -33,8 +33,9 @@
 # nothing reads for correctness -- Metal AOT rejects aliasing with a
 # write target, so it cannot be reused for any of the output buffers.
 
-from std.gpu.host import DeviceContext, DeviceBuffer, HostBuffer
-from std.gpu import block_dim, block_idx, thread_idx, barrier, global_idx
+from max.gpu.host import DeviceContext, DeviceBuffer, HostBuffer
+from std.gpu import block_dim, block_idx, thread_idx, global_idx
+from max.gpu.sync import barrier
 from std.gpu.globals import MAX_THREADS_PER_BLOCK_METADATA
 from std.collections import List
 from std.memory import UnsafePointer, unsafe_memcpy
@@ -166,8 +167,8 @@ def _parse_lean(
         d_structural.unsafe_ptr(),
         d_open_close.unsafe_ptr(),
         d_quote_dummy.unsafe_ptr(),
-        UInt32(size),
-        UInt32(total_padded_32),
+        UInt32(UInt(size)),
+        UInt32(UInt(total_padded_32)),
         grid_dim=num_blocks,
         block_dim=BLOCK_SIZE_OPT,
     )
